# 8321591: (fs) Improve String -> Path conversion performance (win)