@dfuch 2025-04-03
I skimmed through the changes and they look reasonable. I assume there are tests that already cover path normalization on windows?