@bplb 2025-04-03
> I assume there are tests that already cover path normalization on windows?

`Path.of` normalizes by default. There are a lot of cases tested here:

https://github.com/openjdk/jdk/blob/master/test/jdk/java/nio/file/Path/PathOps.java