@bplb 2025-04-03
> I note that the JBS description says "Path.of(URI)" but the benchmark is Path.of(String).

Correct, hence my comments after the table above about other opportunities in `WindowsUriSupport.fromUri`. If there are none, then perhaps the issue and PR summaries should be changed accordingly.