@bplb 2025-04-03
> I assume there are tests that already cover path normalization on windows?

Also, this changes passes JDK tiers 1-3 CI tests.