@RogerRiggs 2025-04-03
I would encourage going further with the refactoring folding two classes into one.
WindowsPathParser is all static methods but with a better name could be a record that is the parsed WindowsPath.
The parsing methods could be either static or overloaded constructors with the input string and normalization flag, invoking the canonical constructor. YMMV.