@bplb 2025-04-04
> I would encourage going further with the refactoring folding two classes into one.

Sounds like a good idea. I will investigate. Thanks.