@AlanBateman 2025-04-07
> Correct, hence my comments after the table above about other opportunities in `WindowsUriSupport.fromUri`. If there are none, then perhaps the issue and PR summaries should be changed accordingly.

I think we should rename, otherwise the commit log will be a bit confusing.