@bplb 2025-04-07
> I think we should rename, otherwise the commit log will be a bit confusing.

Will do. I thinks it would also be worth filing two more related issues: one for the suggested refactoring of `WindowsPathParser`, and another to investigate improving the performance of other portions of the `URI` -> `Path` conversion.